+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "merkle-tree/*",
     "programs/*",
     "state",
+    "test-utils",
     "utils",
     "verifier-sdk",
     "xtask",
