+++ test-utils/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "light-test-utils"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anchor-lang = "0.29.0"
+anyhow = "1.0"
+ark-ff = "0.4"
+solana-program-test = "1.17.4"
+solana-sdk = "1.17.4"
+thiserror = "1.0"
+
+[dev-dependencies]
+rand = "0.8"
